Update ch01.asciidoc
Updated the merchant example and put it into the right place.  Thanks @RicardoM17 for your contributions!   I've replaced the names, as I think the Alice / Bob notation brings some baggage with it that we probably don't need.  (Or if we do want to use it, we should re-apply the names in correct order at the end of the writing process.)