Update ch01.asciidoc
Started to fill out the Getting your First Bitcoin on the Lightning Network.   While custodial solutions are not ideal, they may be a good place for some users to start (given that smaller amounts of money involved) - I think it would be dis-service to the reader not to mention them (and highlight their risks).