Added few extra topics
@renepickhardt  Nice work on the outline. Here are the changes that I have done. Would appreciate your feedback. I have added a couple of topics like `nLocktime`, `CLTV`, `CSV`, and `nSequence` to the "Review of Bitcoin" section as I feel they form the foundation of the transactions in Lightning Network. Also, I have added the topic "Need for a payment receipt" in the "Motivation of Lightning Network" section as invoice payments gives us a cryptographic proof of payment that is generally lacking in a normal on-chain transaction. 

I changed the name of `Breach Remedy Transaction` to `Penalty transaction` as the BRT term does not appear anywhere outside Joseph Poon's whitepaper. I think you should also rethink the RSMC term as that term also does not appear in the BOLTs and outside the whitepaper.

I think it would be worthwhile to add scriptless scripts and benefits Taproot can provide to Lightning Network in BOLT 1.1 /2.0 section. With Taproot, we would be able to do wonderful things with Lightning onion and as the book gets published in 2H20, we would be might be much closer to Taproot implementation on mainchain.