CH1 minor edits and suggestions
The oxford comma, used before the "and" at the end of a list, is standard in american english.

I'm not sure specifically what an "Authoritative" Bitcoin node is, but I feel like it would be implied with just "Bitcoin node" which would be less confusing for the reader.

At times the Lightning Network is framed as blockchain agnostic, but then most of the time it's assumed we're talking about a Bitcoin LN node. For consistency and brevity, I feel like it's better to mention at the start that the LN is blockchain agnostic, but from then on assume we're using Bitcoin.

There was some inconsistency for capitalization of "Lightning Wallet". After it's introduction, I went with lowercase, but I think "LN wallet" would also be an option, consistent with "LN node"