mainly drew a clearer distingtion between payments and transactions
transactions are what is happening on chain and technically speaking a data structure to be placed in the blocks of the bitcoin blockchain. While on the lightning network bitcoin transactions are being modified constantly and outside of the bitcoin network when making a payment we don't do transactions. 
At that point it is not important to understand the technical differences but I think we should already stick to the terminology that we will use later and throughout the book. 

Also added two sentences before the second sentence as I felt we went to the technical part / defintion too quickly.