change wallet categories into wallet quadrant
The original categorization was not perfect. It was not ideal. It was sort of misleading. It sort of implies that a "Full Node & Wallet" is not a "Non-Custodial Wallet" because they are 2 different categories. It also states that a "Non-Custodial Wallet" uses "3rd-party node" for LN Node which is not the true definition.Things were getting mixed up here and we are not communicating a clear message.

The term "Custodial Wallet" is independent of which LN Node is used.
We should rethink this table and categorization.

I wrote an Issue about it #176 .

The Quadrant seems to be more useful and certainly more accurate.

Then I explained the terms "Lightweight" and "Node" because these terms were suddenly used, out of nowhere with no explanation.