Section: Hardware Requirements for Lightning Nodes
Took a stab at writing about the importance of hardware, the kinds of hardware nodes, and what's needed. I decided not to mention particular Node projects as this can easily change. Although to note: we could mention MyNodeBTC, Raspiblitz, and Casa Node

Maybe this section should go higher up in this chapter.

Note: this commit is a response to my previous issue here:

https://github.com/lnbook/lnbook/issues/143