Routing - Creating a Network of payment channels (Language Edits)
Edited the entire section heading above for clarity and ease of understanding. Tried to make parts more concise, but also added a lot of explanatory intermediary text. Standardized "golden"/"gold" coins as "gold". Ultimately, could use another couple of pass throughs to make it read better.

This is a reworking of #370, which has been scrapped