Update 03_how_ln_works.asciidoc
You are right [Umar](@bolatovumar) , I have not found a single source on the web with _two-of-two_, _2-of-2_ seems to be the standard. I edited my pull request, so now the changes do not include 2-of-2 modification. I find the issue of '2-of-2' vs 'two-of-two' interesting and may soon create a question on [English Stack Exchange](https://english.stackexchange.com/) in order to get their opinion.