ch-channel-graph: initial draft of channel graph chapter
In this commit, we present an initial version of the chapter on the
channel graph itself, how to bootstrap a new peer to the network, how to
sync it, how to validate it, and how to maintain it once bootstrapped.

The current version of this chapter covers the following BOLTs: 07, 10.
BOLTs 09, 08, and 01 are alluded to along the way.

This is marked as a draft commit, as I still need to fill out the final
portion that covers how the current gossip network works, and the
protocol used to sync the channel graph based on block heights of
channels.

Due to the current length, we may want to consider eventually breaking
this chapter up, however I'm focusing on content first.