ch-channel-graph: fill out latter portion of chapter
In this commit, we pick up where we left off last and describe the
semantics of the `channel_update` message including how to validate it
and the significance of the message. We also describe in detail the usage
of the `channel_update` message within the routing network, which
includes an explanation of the fee structure, and the other fields
within the message itself.

To conclude the chapter, we describe the process used to _sync_ the
channel graph in the first place, as well as how peers maintain the
graph once it has been fully synced.