wire_protocol+upgrades: initial version of chapter
In this commit, we add the initial version of a chapter that explores
the framing of the wire protocol including an exploration of the various
upgrade mechanisms available to extend the LN. Portions of this chapters
can be used to fill out other chapters such as the funding flow. The aim
of this chapter was to provide a single point in the book that readers
can go to in order to get a high level understanding of the wire
protocol of the LN. This chapter doesn't go into low level protocol
flows involving the messages, as that's to be left for the chapters that
dive deeper into the content.