Update preface.asciidoc
Addresses don't use bullet points. 

(I am new to this, so I'm not sure it two spaces at the end of a line is the best way to indicate a line break.)  I'm planning on reading the whole book - I have quite a lot of experience of editing for publication.  Expect more from me.