Update 02_getting_started.asciidoc
Sorry for the multiple patches for the introduction.  Getting my head around this.
I've made some changes in capitalization (using the Glossary as a reference) and punctuation. I've also tidied up one or two phrases that were unclear/awkward.

Note this from the guidelines:
Headings should be all lower case, except for first letter, proper nouns and acronyms. "An introduction to the Lightning Network", "Explaining the physics of fulgurites" etc.

Hopefully, I've corrected all the headings.