Update 05_node_operations.asciidoc
Capitalization, formatting and minor rephrasing of one or two clauses to make them clearer. 

I'm working through the whole book, chapter by chapter, and so am trying to increase consistency.  For example, when terms are introduced, they should have the acronym in parenthesis the first time, not the second time.