ch-brontide: add initial draft of the brontide chapter
In this commit, we add the initial draft of the chapter on Brontide
which describes the encrypted messaging transport used in the Lightning
Network. At the time of this commit, this chapter is ~90% complete.

After the initial background, motivation, and context, the actual
handshake process is rather mechanical, so I've opted to lift by text
from the original BOLT with minor modifications. However the change from
the background/intro to the description of the protocol is rather stark
as it shifts to more of a technically oriented prose. I plan to revisit
the latter portion of the chapter to provide more insight that is
presently checked in.