ch-invoices: add initial chapter on BOLT 11 invoices
In this commit, we add a chapter that describes the BOLT 11 invoice
format at a high-level as well as the set of specific fields and base-5
encoding used within the data segment. The only thing this chapter is
missing now is likely some test vectors, and possibly a bit more
mechanical information that an implantation must observe when
encoding/decoding an invoice.