ch-07: use a private IP addr for example node IP
As is the chapter uses 144.76.235.20 which has a PTR DNS record to
rene-pickhardt.de. Rather than include possibly private information
relating to the authors of the book, we opt to use an IP address from a
known private range. This doesn't take away from the example, but lets
us include less personally identifying information in the book.