ch-07: mention asymmetry w.r.t interpretation of to_self_delay field
A party can't pick its to_self_delay (the CSV delay) itself, as that
would mean they can set it to effectively zero and be able to breach at
will. Instead, both sides propose a CSV delay for the other, effectively
enforcing time delay that allows them to act in the face of a breach.