Update 04_node_client.asciidoc (#678)
I've updated the headings to match the recommendation in the style guide, namely that 'Headings should be all lower case, except for first letter, proper nouns and acronyms. "An introduction to the Lightning Network", "Explaining the physics of fulgurites" etc.'

I've also capitalized Docker where it refers to the name of the software (not to commands).  I spent a lot of time on this to get it right.

I also removed the plusses (+) from two examples that referred to the Docker, the software, and not to Docker commands as this retained consistency with the rest of the chapter.

There was a spelling mistake and some other typos.