ch-10: add initial explanation of session key randomization
In this commit, we include an initial explanation of the session key
re-randomization scheme that makes the Sphinx packet format so compact.
Alice is able to generate all the necessary session keys up front by
using an initial session key, and doing an EC multiplication of that key
and a blinding factor derived from information of the last hop.

I wanted to use sub-scripts for a lot of the notation here, by the
asciidoc renderer I use didn't seem to be able to render them....

I think I want to make another pass and blend in the initial explanation
in the high level style of the initial ECDH explanation.