wire_protocol+appendix: split portion of wire appendix into new chapter
In this commit, we split out the potion of the wire protocol appendix
that doesn't enumerate all the messages types into a new chapter. This
new chapter describes the framing of the wire protocol, the TLV scheme,
feature bits, and finally how the protocol is upgraded in theory and in
practice.