Update 03_how_ln_works.asciidoc
I think Bob don't retaliates by sending the true commitment transaction (true state of the channel). If the cheat is discovered he is able to claim all the funds by executing a punishment transaction. Submitting the true state would be a double spend and would lead in the worst case to Bob getting less. I am not a big Lightning expert, though.