Correct encoding scheme for obfuscated state hint
The upper 64 bits of nSequence are not used to encode the state hint, as
there are only 32 bits in nSequence.  And the upper 8 bits are not used
at all, since that could inadvertently enable a relative timelock by
unsetting bit 31 (see BIP 68).

From BOLT 3:
```
Commitment Transaction
----------------------
locktime: upper 8 bits are 0x20, lower 24 bits are the lower 24 bits of the obscured commitment number
txin[0] sequence: upper 8 bits are 0x80, lower 24 bits are upper 24 bits of the obscured commitment number
```