Corrected error regarding Loop services and inbound
Acquiring inbound liquidity is done through Loop-Out, not Loop-In. Loop-in is when you send the service on-chain bitcoin and then they pay you on lightning, thus increasing your outbound. Loop-out is when you pay on lightning (shifting liquidity from outbound to inbound) and receive funds back on-chain minus the service, etc. fees to increase inbound. The section is talking about how to acquire inbound, not outbound. The PR is meant to reflect the correct application of the topic.